APP: FIX LINT ISSUES - Add userInput dammmy data and migrate the data from estimator file - add console.log() rule in eslintrc file
