+++ src/dammy data/userInput.js
@@ -1,0 +1,13 @@
+const userInput = {
+  name: "Africa",
+  avgAge: 19.7,
+  avgDailyIncomeInUSD: 5,
+  avgDailyIncomePopulation: 0.71,
+  periodType: "days",
+  timeToElapse: 58,
+  reportedCases: 674,
+  population: 43422705,
+  totalHospitalBeds: 1380614
+};
+
+export default userInput;
