+++ src/estimator.js
@@ -1,16 +1,5 @@
 import generateData from "./helpers/generateData";
-
-const userInput = {
-  name: "Africa",
-  avgAge: 19.7,
-  avgDailyIncomeInUSD: 5,
-  avgDailyIncomePopulation: 0.71,
-  periodType: "days",
-  timeToElapse: 58,
-  reportedCases: 674,
-  population: 43422705,
-  totalHospitalBeds: 1380614
-};
+import userInput from "./dammy data/userInput";
 
 const covid19ImpactEstimator = (userData) => {
   const data = generateData(userData);
@@ -29,6 +18,6 @@ const covid19ImpactEstimator = (userData) => {
   return result;
 };
 
-// console.log(covid19ImpactEstimator(userInput));
+console.log(covid19ImpactEstimator(userInput));
 
 export default covid19ImpactEstimator;
